feat: automatically remove users from renewal emails
