+++ sale_actions/config.template.toml
@@ -4,6 +4,7 @@ check_delay = 10
 [email]
 base_url="https://connect.mailerlite.com/api/subscribers"
 api_key = "xxx"
+ar_group_id = "xxx"
 
 [database]
 name = "goerli"
