+++ sale_actions/src/main.rs
@@ -28,8 +28,8 @@ async fn main() {
     }
 
     loop {
-        processing::process_purchase_data(&conf, &db, &logger).await;
-        processing::process_auto_renew_data(&conf, &db, &logger).await;
+        processing::purchases::process_data(&conf, &db, &logger).await;
+        processing::renewal::process_data(&conf, &db, &logger).await;
         sleep(Duration::from_secs(conf.general.check_delay)).await; // Sleep for 60 seconds before repeating
     }
 }
